requirements.txt
streamlit==1.20.0
python-dotenv==1.0.0
Pillow==10.0.0
google-generativeai==0.1.0
